mw-api-int: Double envoy concurrency
While migrating RESTbase backend traffic, we observed larger connection
establishment times and errors than on bare metal and other k8s service
deployments that don't have this concurrency restriction.

As the concurrency number was experimental, increasing it to figure out
the right value on mw-api-int first, and possibly to all mw-on-k8s
afterwards if results are conclusive.

Bug: T358213
Bug: T354532
Change-Id: I441fbd8ba614f7910759acf95969a22706b91dc3