superset: ensure role list returned by OIDC server is a list of strings
This fixes the bug reported by Neil Shah-Quinn, in which he was only
assigned the Alpha superset role (the default one), when he should have
had the sql_lab role as well.

This is due the CAS returning `role_keys: wmf` instead of `role_keys: [wmf]`
in the case of a user pertaining to a single group.

As the method iterating over the user LDAP groups to assign the corresponding
Superset roles was expecting a list of string, it ended up iterating over the
*characters* in the string role, instead of iterating over a list containing
a single role, thus not matching anything.

Bug: T361894
Change-Id: I993bdad1265ac044b33f223624ff4727430a5fe6