cirrus: Tune resource usage of consumer
We've been having some OOM issues since deploying the updated
elasticsearch sink. Resolve by reducing the maximum size of bulk
requests we allow processing, along with increasing the number of
taskmanagers to spread the same work to more instances. The fetch
queue capacity was reduced to match the increase in taskmanagers,
the capacity is per-taskmanager and we want to keep a limit on
the total capacity. While testing this configuration we have
seen significant throttling, allow more cpu usage as well.

Change-Id: I24640a207ec628f06ea10e2207ab571b4989bc2d