kask: allow to configure tls options
At the moment kask's _config template overrides cassandra->tls
settings with the ca bundle path, regardless of what is
set in other places like values.yaml. In the future we will
move Cassandra's TLS certs to PKI, so we'll also need to specify
a different CA bundle path (at the moment the only option is
the one to use the hardcoded fixed path).

* Move the declaration of the CA bundle path to use to be
  an explicit yaml setting.
* Fix helmfile.d configs that use the kask chart accordingly.
* Fix enable_host_verification setting for session-store
  staging.

Bug: T352647
Change-Id: Icc049aeca36dac5d90ee9180010f58915c0506ca