knative-serving: move net_istio configs to a dict
The idea is to be able to easily and selectively override
values in staging without the need of copy/paste
the whole existing config (since we use arrays this is
required since we can't merge values from multiple sources).

While at it, move the knative's net_istio specific config in
staging to a more appropriate place.
Finally refactor the current configs to adopt the new dict scheme.

Bug: T353622
Change-Id: Id232dfe5943056558579f65fd9007214d087516f