admin_ng: move Istio configs to mw-api-int-ro for ml-serve
The ServiceOps team is migrating every usage of api-ro.discovery.wmnet
to mw-api-int-ro.discovery.wmnet.
As part of the move, I discovered a bug in our current Istio config
for the MW API, explained in T353622#9723070

To do the migration, we need to:
- Avoid IPv6 lookups for coredns (to bypass an Istio bug)
- Add a specific list of external endpoints (*.wikipedia.org, etc..)
  as ServiceEntry. This wasn't needed before due to an Istio bug,
  allowing calls to $domain:80 without $domain being listed in
  a ServiceEntry.
- Update the Virtual Service config to use mw-api-int-ro instead
  of api-ro.
- Drop the api-ro configs.
- Support both proxy use cases: explicit (like forcing the aiohttp's
  calls to a discovery endpoint) or implicit/transparent (namely
  using http://en.wikipedia.org directly for example).

These changes will be applied to every DC depooling it from traffic
beforehand, since we'll also need to deploy the updated configs
for the model servers.

Bug: T353622,T362316
Change-Id: I28ef52a8df0e1ad97a9445367c89d73f5a07e9b7