ml-services: tweak reference to ML Cassandra clusters
In the course of change 1020194 it became clear that aggregating the
clusters between DCs there is less useful than doing it in a deployment
chart. So instead we use the generated names here instead of hardcoding
every cluster name in the other change.

This depends on the above-mentioned changed to be merged before it
creates the right network policies.

Bug: T360428
Change-Id: I620a374c44eb7b6195d66e2fa49f342970403b57