kserve-inference: allow transparent proxy mode for revscoring isvcs
The WIKI_URL env variable is mandatory for revscoring isvcs.
We want to be able to automatically set it when we use
the transparent proxy functionality, so this change adds
a specific code path to auto-generate the WIKI_URL env
variable in case it is not set by the operator in either
base_env or custom_env. The change adds also two new
tunables to change the port and the protocol (in theory
not needed but better to have them with some sane defaults).

Bug: T353622
Change-Id: I07db96637c59b33f0302df184f8a67792e8edeb1