cirrus updater: Reduce saneitizer capacity and fix backfill arguments
The saneitizer capacity was arbtrarily set to 100. Runtime usage looks
to be quite steady, using 5-10% of this capacity. Bring the configured
capacity down to 20, which should result in 25-50% of capacity used
during normal operations.

Additionally a recent change to the saneitizer, so that it doesn't
change the shape of the graph based on configuration, has required an
additional parameter to be added to the backfill jobs.

Change-Id: If89186292bd104b7b33078266e3c6dc6d3ed403a