cirrus updater: Enable codfw consumer
This is the first full production deployment of the cirrus updater. It
copies over the configuration we have validated on cloudelastic and
will start performing the same updates, but now against the codfw
cluster.

This is initially constrained to a smaller set of wikis that adds up
to ~25% of the full production volume. If everything is looking happy
here we will move it to 100% in a few days.

Bug: T363475
Change-Id: Ib971c6a311f599d7befc21c2147772d360b07a29