Make base.certificates compatible with chart modules and scaffold
Make the certificate module compatible to module standard (in terms
of service and object names, annotations etc.)
Align the certificate module with the mesh.certificate modules API

WARNING: Backwards incompatible change!
certificate 1.0.0 used to include the certificates objects name
into the SAN's of the certificate which made it require manual
configuration to work with default scaffolding and modules.
This version aligns SAN's so that they are compatible with default
scaffolding and modules by default.

Bug: T362310
Bug: T258934
Change-Id: I04c405c6bab364dd99d13195ee16a0c55ed06b13