ml-services: further tune autoscaling for editquality-reverted
Like commit 3aebc3ec51383055202ddc955411c0f6c153fff7, this attempts
to address the increased latencies in get_revscoring_extractor_cache()
by decreasing the target rps per pod from 10 to 5.

We only adjust editquality-reverted in this change, and if it does help,
we may adapt -damaging and -goodfaith in a similar way.

Bug: T363336
Change-Id: Ifd2c82a3e263b4b426872594e6c01411ed11df11