cirrus: Expand codfw to serve writes to all public wikis
This has been running at ~25% capacity for over a week now and things
are looking happy. Shift the rest of the public wikis over.

This removes the wikiid's filter, matching how the producer is
configured. This does not include private wikis, as private wikis
are not (yet) included in the event streams.

Bug: T363475
Change-Id: I4e93c57d4512f4b23d7ad7faf039455291078ffc