Deploy the ceph-csi-rbd chart to dse-k8s with default values
We are evaluating the ceph-csi-rbd chart to see whether or not it is
suitable for use on the dse-k8s cluster.

The chart has been added and customised in the previous two commits.
This commit adds the required configuration to the admin_ng service to
deploy it to the dse-k8s cluster only. We deliberately do not override
any of the values in this commit, in order to be able to evaluate what
the default configuration would attempt to do.

In the helmfile we add a calico networkpolicy to allow the provisioner
components to communicate with the Kubernetes API.

This commit should only be merged along with the subsequent commit,
where the values are customised for our environment.

Bug: T364472
Change-Id: Ifb8bb79055f79cce64ff4766313cd5a08cd0f5d1