services: move Tegola's Swift config in staging to local envoy proxy
Attempt to use the local TLS proxy (envoy) in Tegola staging
as opposed to call the Swift Thanos S3 endpoint directly.
Hopefully if this works Tegola will be able to contact
Thanos Swift with different TLS settings, not incurring
in high cpu usage (for example, when we tried to move
Thanos Swift to CFSSL/PKI).

Bug: T344324
Change-Id: Ib11149235f73fad7d0f1da3f0b317d28e41eaf0d