ml-services: Update Docker image for nllb-gpu
We are currently debugging some GPU issues in staging, where
we already use the new entrypoint.sh scheme (that creates in
turn the python model server), meanwhile we still don't use it
for nllb-gpu. Updating the Docker image to make sure this is
not the culprit of the problem (unlikely but better safe and
reduce the variables on the table).

Bug: T362984
Change-Id: I0c7d1538e272086ec264a5de54736b715f6901c7