admin_ng: Add Cassandra ServiceEntry and VS for LiftWing
The original patch (1029139) had to be reverted since combining
hostnames with opaque TCP connections does not work. This change instead
uses TLS as protocol, plain hostnames, and DNS resolution. This should
allow Istio to see the TLS part of the connection and after TLS
negotiation stop caring about the specifics.

Bug: T360428
Change-Id: Ic8a538d605bc7bda2fdd391439fd8f59588b33db