otelcol: tweak rollout params
A rolling restart takes longer than 10 minutes, currently
causing Helm to revert it to the previous state after it
spends 20 minutes doing all the work.  Bump timeout to an hour.

Also we only restart at most 1 at a time, which makes it take 20
minutes to restart a fleet of ~190 pods.  This number will only
increase in the future.  Run more restarts at once.

Bug: T363407
Change-Id: Ib33f6da64df9855f922167c56bf02cd68763ff76