Add a values file for the ceph-csi plugin on dse-k8s-eqiad
This values file will be used to configure the cephcsi plugin deployment
on the dse-k8s cluster in eqiad.

It contains details of the ceph mon servers, as well as the fsid of the
cluster, plus our custom images for each of the required containers.

We have decided to disable the snapshotter functionality, since we do
not currently require it and it needs further CRDs to be deployed.

Bug: T327259
Bug: T364472
Change-Id: I80eff44043a8e27ede8b4dbbad87b2b867e6d840