charts/kserve: Switch to using k8s service targets for NP
Instead of using the hardcoded IP/Mask entries to allow kserve to talk
to the k8s API (and vice versa), use the service entries.

For now, add two crd.projectcalico.org/v1 NetworkPolicies, but keep the
old policies intact.

Note: I am totally unsure if this is the right way to do it. For one
thing, I am not sure the NS is correct, secondly, I don't know if the
service entries for the k8s API are autogenerated or that should be part
of this change. Unfortunately, the documentation for k8s 1.23 is gone
from the official website.

Bug: T365479
Change-Id: I9a6cff600a2366103695cc826b6e17d645643009