mesh.configuration: Add support for rate limiting
This adds support for rate limiting using the ratelimit service to the
mesh module. It allows services to opt in to being rate limited by
adding a (empty) map with the listeners name to the
mesh.ratelimit_listeners structure. Rate limiting then happens in the
local envoy on the "client"/"caller" side.

The rate limit domain (needs to be configured in the ratelimit service
as well) will default to the name of the listener and can be overridden
by the "domain" key in the ratelimit_listeners map.

Alternatively to client-side configuration a listener can enable rate
limitng for all clients by specifying a "ratelimit" map. Re-deployment
of all service-mesh "clients" is required currently to make this
effective.

As of now only rate limiting by user-agent is supported.

Bug: T362310
Change-Id: I180bdd09e2cfc35f9110c23725f00044f898cf3c