Migrate AQS2 services and image-suggestions to calico network policies
We have updated the aqs-http-gateway chart to support the calico network
policies for the support of external services.

This commit updates the chart as well as all of the deployments that use
that chart. We only add cassandra to the external_services and not
druid, since the druid public cluster is only ever accessed via the
service IP, not the IPs of the worker nodes.

Bug: T359423
Bug: T364533
Change-Id: I18e9579b84eb77dbc61558d44f151ef3c005c894