sessionstore: update certs in advance of expiry
Sessionstore certs are due to expire very soon. This service is unique
and old enough that it does not use any of the standard cert
management patterns on k8s, and so requires us to generate keys via
puppet scripts.

Testing certs in staging has gone okay so far, and the key has been
regenerated in puppet private. This change will need to be deployed
alongside the relevant private key change in puppet private's
deployment_server module and not before.

Bug: T363996
Change-Id: I17f9e33a436fdb47e9fb4eddde913f518b0a57d9