api-gateway: add normalise_paths option, enable in api-gateway
Currently envoy defaults to not normalising URL paths, despite being
the RFC recommended action to take. This is currently breaking some
interactions with LiftWing where clients are using normalised
paths. Enable normalisation for the API gateway along with adding an
option to toggle it (which may be needed on the REST gateway in future
also)

Bug: T365439
Change-Id: Ia5530531c83b327a26946d6b08e07247265e731b