otelcol: deploy k8sattributes processor
This associates many fields of k8s metadata with incoming traces,
matching based upon the pod IP address of the reporter.

For now, we hardcode in the IP addresses of the k8s API servers so
we can create a network policy for them.  (See T365855.)

Also, bump the memory requirements, since the k8sattributes processor
was otherwise causing the memory_limiter processor to refuse traces.

Bug: T363407
Bug: T365855
Change-Id: If1ccc1c7eef120124657ac2695d22d87b963c898