datahub-next: make sure subcharts get the environment default values
What
----
Make sure to inject the environment default values (define in `general-dse-k8s-eqiad.yaml`)
into the subchart rendering context.

Why
---
When we rendered and deployed datahub-next to dse-k8s-eqiad, we realized that we were missing
the kafka-test brokers from our network policies, which was weird because the broker IPs appear
in the environment values.
However, due to how the datahub chart is structured (a main chart with a subchart for each component)
the global env values were not injected in the subchart `.Values` object, due to the
```
- "/etc/helmfile-defaults/general-{{ .Environment.Name }}.yaml"
```
line missing from the environment values.

My theory is this: we're knee deep into "YAML overriding bonkers territory", meaning that
by specifying `/etc/helmfile-defaults/general-{{ .Environment.Name }}.yaml"` as part of the
environment values, `helmfile` injects it in the release values, and all the subcharts.

Bug: T361185
Change-Id: I57aa3d90af0aa247730ab2e7e98fe0bc3ad4169c