Configure datahub-gms-next not to wait for upgrade before starting
The env variable BOOTSTRAP_SYSTEM_UPDATE_WAIT_FOR_SYSTEM_UPDATE is required so that
the migration job can contact the GMS and initiate the process to upgrade.
We are not running an upgrade at this time so we shall be disabling this.

This also Adds the environment variable AUTH_RESOURCES_DIR to the
datahub deployments, in order for the datahub-gms service to start as is done
in production.

Bug: T361185
Bug: T363300
Change-Id: I10a4ccd72496df96b54f5e2df3cee32ed820ed11