datahub-next: flip env var to true to support internal registry
Deploying datahub-next failed with the error reported in
https://github.com/acryldata/datahub-helm/pull/358

If we confirm that this patch fixes the deployment, we will change the
hardcoded value of this environment variable in the chart. For now,
we set it as an override to the datahub-next deployment, to avoid
causing issues with the production release.

Bug: T363461
Change-Id: I2707467c98a8fe1a50ba2e0d3d0bd072c98648c7