mediawiki: allow passing variables to php-fpm
When running php-fpm, it is useful to be able to pass environment
variables we either inject or are available in the pod. To this end, we
have created change  I959d72588a1 which introduces a /env directory
where we can place files that will be sourced by the php-fpm container.

This change introduces a new key in the values.yaml file called php.envvars
that allows to pass both a list of environment variables with included values
to the container, and to pass those up to the php execution environment, together
with other variables that are available in the pod.

Change-Id: Ie879fe60e3c04fd46f21ab7343f36d82e1e46472