mw-debug: start using php.envvars, expose statsd-exporter
We want to expose the statsd-exporter service IP to mediawiki.
To this end, we add the php.envvars variable to the mw-debug
deployment, adding the STATDS_EXPORTER_PROMETHEUS_SERVICE_HOST
variable that kubernetes provides, making it available to php.

By populating the php.envvars variable, we override the default
php-fpm configuration to use the environment variables that are
set in the container. So we need to redefine the MCROUTER_SERVER env
variable, where the value was overriden.

Bug: T365265
Change-Id: I5da1e3caf0541c7dd60535bc11f0c5f45257e334