proton: drop replicas from 12 to 10
It turns out that, after the bump to 12 replicas in
I1f976131f2d8f1667efc00e005011a554fde88a2, we really only ever got 10
given the default quotas.

Let's lock in 10 as our desired replica count and (in the next patch)
bump the quotas a bit to ensure there's plenty of room for deployments.

Bug: T362978
Change-Id: I755d9079542a29ee74ba2cead986feb86d809b18