Allow running CI in a container when using rootless podman
When running containers as rootless podman, the root user in the container
maps to the current user on the host. This means that, for instance, all files
owned by the user in the host will be owned by root in the container.

So, when the environment variable ROOTLESS_PODMAN is set to 1, run the
container as "root".

Change-Id: I72821a14d41b13112753569fbfafdc06fb7430bc