datahub: don't use an ES index prefix for production
We added an index prefix for staging because we didn't want both releases
to use the same indices. However, if we did release that change, we would
end up with an empty production datahub instance. We'd have needed to run
reindexing jobs, as well as to clean all old indices. We can avoid all of
that by keeping the old indices.

Change-Id: I31aaa82a125d1e14d3e136fa64c3870ae84613be