mediawiki: restart deployment on env changes
This commit updates the mediawiki chart to restart the deployment
when the environment variables change. This is done by adding an
annotation to the deployment.

This is necessary because not every env variable is explicitly declared
for the container, but rather they can be provided by kubernetes
so adding/removing it from the configmap would not trigger a redeploy.

Change-Id: Iaec41cf74176cbaafd0b293a670954026facc98f