mediawiki: add securityContext to all containers (attempt 2)
Alerts and dashboards have been updated to support the local_service ->
LOCAL_<release name> rename, so we're ready to try again.

Note: Unlike the first patch in the prior series, this updates both
mw-debug and canary releases (on all deployments supporting them), as we
have higher confidence that this works as expected.

Original commit message:

Adds the restricted-PSS compatible restrictedSecurityContext to all
containers, with the exception of the php app container, which needs
SYS_PTRACE.

For that, a nearly-identical securityContext is defined inline, but with
capabilities.add: [SYS_PTRACE].

Other than SYS_PTRACE, the only notable difference between the existing
mediawiki and restricted PSPs is the handling of hostPath volumes, which
has no analogous support under PSS (and was only validating anyway), and
makes restrictedSecurityContext a reasonable starting point.

This of course does not make mediawiki compatible with restricted PSS,
but is about as close as we can get.

Supporting module updates:
* Patch version bump on base.helper for restrictedSecurityContext.
* Patch version bump on base.statsd, cache.mcrouter, and mesh.deployment
  for securityContext on statsd exporter, mcrouter, and envoy.
* Minor version bump on mesh.configuration (1.6 -> 1.8) to unblock the
  above.

Why the redundant chart field when pinning the version? This appears to
be necessary in order to convince CI to generated the expected diffs.

Bug: T362978
Change-Id: I75b4c97319e4f2939a607afab84edb1133cf6805