mediawiki: add securityContext to all containers
Adds the restricted-PSS compatible restrictedSecurityContext to all
containers, with the exception of the php app container, which needs
SYS_PTRACE.

For that, a nearly-identical securityContext is defined inline, but with
capabilities.add: [SYS_PTRACE].

Other than SYS_PTRACE, the only notable difference between the existing
mediawiki and restricted PSPs is the handling of hostPath volumes, which
has no analogous support under PSS (and was only validating anyway), and
makes restrictedSecurityContext a reasonable starting point.

This of course does not make mediawiki compatible with restricted PSS,
but is about as close as we can get.

Note: Other than mw-debug, all mediawiki releases are pinned to the
prior chart version, in order to control / proctor the rollout.

Supporting module updates:
* Patch version bump on base.helper for restrictedSecurityContext.
* Patch version bump on base.statsd, cache.mcrouter, and mesh.deployment
  for securityContext on statsd exporter, mcrouter, and envoy.
* Minor version bump on mesh.configuration (1.6 -> 1.8) to unblock the
  above.

Bug: T362978
Change-Id: I405a16c94a6a98f09c8ec9b952c3abd29f16f9d3