mediawiki-dev: add securityContext to all containers
This chart is for local development only. As such, it's not strictly
necessary to make it compatible with restricted PSS (nor would it be
possible, given the way certain containers are configured).

I've added restrictedSecurityContext to all containers except the init
container that must run as root. I've also added a config value that
controls whether these used, defaulted to false (and documented) and a
fixture that flips it true for chart tests.

From a quick scan of both debugging extensions, I don't think this
should prove problematic. However, offering an escape hatch via the
config value seems reasonable.

Bug: T362978
Change-Id: I43d96e937d4beccfafdc29f55de273f653dd2c90