cloudnative-pg: adjust RBAC management by scoping it to PG cluster namespaces
As per akosiaris' review (https://phabricator.wikimedia.org/T364797#9913576),
we disable RBAC management within the chart, as it was viewed much too broad.

We will manage RBAC namespace by namespace instead, by specifying the namespaces
to be watched, and will create `Role` and `RoleBinding` resources in each of
these namespaces.

However, some resources, such as nodes, namespaces and webhooks are not scoped
to a given namespace, meaning we wtill need to define a `ClusterRole` and associated
`ClusterRoleBinding`, with a muich slimmed-down permission list.

We have also evaluated the roles that will be created by the operator
here https://phabricator.wikimedia.org/T364797#9920892, as per the ns-scoped RW
permissions on roles and rolebindings.

Bug: T364797
Change-Id: I475bb266c166e692374e1f3583cfba8e9568a4f3