admin_ng: Bind to privileged PSP if restricted PSP is disabled
All workloads are required to pass PSP validation in our clusters. So
removing the binding to the restricted PSP is not possible. Instead, we
bind to the privileged PSP, effectively bypassing PSP validations.

As roleRef is an immutable field, the allow-psp RoleBinding is replaced
with a RoleBinding called disable-psp - which also makes it more
obvious.

Bug: T273507
Change-Id: If6b504b68ad2f6c3ca126a88b4cc51f568d341e0