irc.wikimedia.org: Stop sending broadcast events to the old buster nodes
The irc.wikimedia.org CNAME has been failed over to the new bullseye
server(s) irc1002 (and it's failover nodes irc2002). I've forced a reboot
of irc1001 to force all clients to reconnect to irc1002, as such we can
stop sending broadcast events to the old nodes so that they can
eventually be decommissioned.

Bug: T331702
Change-Id: Ibefcbad929a6701522d7a7aad19f35b7ceb18fc6