cephcsi: Run the csi-rbdplugin container as gid 900
We need the node-registrar container to be able to read from and write
to the socket that is created by the csi-rbdplugin container.

In order for this to happen, we can configure `runAsGroup: 900` for the
rbdplugin container. The `fsGroup: 900` should also help to configure
the umask in such a way that the same group can access the socket.

Bug: T327259
Change-Id: I9f4fac8ebffa91d00774c9602cc8d651c41a30fc