cephcsi: Grant elevated privileges to the driver-registrar container
The driver-registrar container is required in order to register the new
CSI plugin with the kubelet. It does this by creating a unix socket in

the hostPath of /var/lib/kubelet/plugin_registry via a volumeMount to
the /registration path.

The driver-registrar also needs to communicate with the socket that
is created by the CSI plugin itself in the hostPath of
/var/lib/kubelet/plugins and mounted to the /csi path.

In order that these two components can communicate between each other
and reconfigure the kubelet, we need both components to be able to run
as root.

We have to specify runAsUser:0 because both images are built with
blubber/kokkuri and so they default to a runuser of 900.

Please see: https://phabricator.wikimedia.org/T327259#9947288 for
further rationale of why we feel that this is required.

Bug: T327259
Change-Id: Ib429e8f754a426ca5c4ca54cff5cd1ba22599e02