knative-serving: remove _example settings shipped with upstream yamls
While trying to update staging, I realized that there was a testing
setting on config-defaults that was applied manually, so I tried
to restore the original _example config but I ended up in:

https://github.com/knative/serving/issues/8211

This seems something very brittle to rely on, the webhook validator
can potentially reject the config if we add new (real) values etc..
It never happened in the past, but I stumbled upon it so I'd vote
for removing the problem entirely to be safe.

Simply stripping the _example settings works, since the checksum
value stored in the annotation is not validated anymore. I think
that it is way better to avoid _example, we have always checked
upstream docs for new settings, it shouldn't be a big problem.
Also those examples can be consulted in the yaml files shipped
by old knative releases on github, so easily retrievable.

Bug: T368359
Change-Id: I558ed01eb44bf230e61fdf14cc8537ac941618a0