cephcsi: Enable the prometheus-liveness container
We had previously disabled the prometheus-liveness container within the
cephcsi provisioner deployment. The reason we did this is that the
cephcsi binary attempts to create /etc/ceph/ceph.conf if it does not
exist. Our containers do not allow this creation, so the
liveness-prometheus container was crashing.

With this change to the chart, we mount /etc/ceph from the existing
configmap and attempt to re-enable the prometheus-liveness container again.

Bug: T327259
Change-Id: Icc6a1370d604fc69a3732d43a694e9df67926b7b