cephcsi: Grant the provisioner access to the ceph userID secret
We had removed cluster-wide privileges for accessing secrets from the
cephcsi provisioner, since we believed that these privileges were only
used when handling encrypted volumes.

It turns out that the provisioner still needs the ability to access the
secret that contains the ceph client userKey.

We can grant this specific privilege without opening up the access any
further, since we know both the namespace and the name of the object
that we need to access. So we can add this to the provisioner's role.

Bug: T327259
Change-Id: I2f7fabf94b9fb521508acd310f92bc0271583821