Facilitate a role swap from an-mariadb1001 to an-mariadb1002
We need to make an-mariadb1002 the active master for the analytics_meta
MariaDB service, in order to allow services to continue functioning
whilst the switch to which an-mariadb1001 is upgraded.

This commit includes changes to the superset and MPIC application
configurations, which will effect these changes. They will need to be
deployed out at the same time as the roles are switched on the MariaDB
servers.

Bug: T365503
Bug: T348977
Change-Id: Ieda7e318025e57fbc8d50324b04f3c6deeb95fe9