shellbox-video: set log level to debug temporarily
What:

Increase shellbox-video level to be verbose for a short period. This
is currently only used for video transcodes on a single wiki
(testwiki). Also don't uppercase all config.public values so as to
maintain the case that php-fpm expects.

Why:

Debugging timeouts between mediawiki and shellbox is proving
challenging, and getting a bit more information about when requests
begin and end would be helpful.

Bug: T356241
Change-Id: I54fd5ca1bd5a485ef6f1e9f908708ecde090a224