Initial commit validating-admission-policies chart
The validating-admission-policies creates policies to replicate the
functionality of Pod Security Standards with a bit more flexibility on
how they apply to namespaces. It's also meant to be expanded with custom
policies.

Apart from the things imported from kyverno's policy repository I've
added missing tests for disallow-proc-mount and fixed the
restrict-volume-types tests:
https://github.com/kyverno/policies/pull/1075
https://github.com/kyverno/policies/pull/1076

Bug: T368251
Change-Id: I7b9af7fd0d0d26ff9de6106b3a37cf74465017a2