growthbook: split chart into 2 charts (frontend/backend)
We need to split the chart into 2, one for the frontend and
one for the backend.
The reason for that is that we need to be able to route different
ingress domains to differen ports, cf
https://docs.growthbook.io/self-host/env#domains-and-ports

The easiest way to perform this domain/port mapping is to manage
different charts, sadly.

Bug: T365839
Change-Id: I016fbf34713aa4ef6273e71abf68e137a6fd84ec