ml-services: override staging rec-api entrypoint
Set the command to run to poetry run guvicorn in order to override the command set in the previous version of the app. Without this the service in staging uses the command + args used for production.
This can be removed when the new service is deployed in prod and the deployment configuration is updated as there wouldn't be the need to override anything and the entrypoint from blubber will be used.

Bug: T371465
Change-Id: I430a4774b2ca42010fcbfe6b8211322ceeb0f599