cloudnative-pg: revert back to chart name = cloudnative-pg
It turns out that the operator code performs filtering on Deployment
resources based on labels, with the hardcoded label `app.kubernetes.io/name:cloudnative-pg`,
and setting this label to any other value causes a runtime crash of
the operator.

To avoid further confusion, we align the chart name with the directory
name that contains it.

Bug: T364797
Change-Id: Iee24797f5290725ab3719285f459434b731a18f4