cloudnative-pg: bump major versiom
We had initially released the chart version 0.21.4, which
is the upstream version, causing any update we make to our own
chat to have a a lesser version than this one, as we decremented
the version back to 0.1.0.

By incrementing the version to 1.0.0, we ensure the current version
gets precedence, and is considered the latest one.

Bug: T364797
Change-Id: I37d0b7c3ca84ec3539e2b69d1c0195ab8049783f