cloudnative-pg: define network policies allowing traffic to and from the k8s API server
The operator has to be able to reach out to the kube API server to list, watch, ... resources.
However, the kube API server also has to reach to the operator service, due to the
admission/mutation webhooks. When certain resources are created, the kube API server
calls the operator service with the resource payload, which is then either validated
or mutated by the operator itself.

Bug: T364797
Change-Id: Ic8b759e6029ace038d0b9a086b0e81561901b6c1