mediawiki: Build sidecars annotation dynamically
Previously the annotation was just a static string, with an optimistic
comment saying we'd keep it up-to-date. That broke with
I996de4d922fcf10a1a3fe557ceee170502dbc724, which switched off the
mcrouter container (so the sidecars controller now doesn't clean up
mwscript pods).

Now instead, check the Values flag enabling each sidecar container (the
same flag used in the container template) and include each container in
the annotation only if it's actually enabled.

Change-Id: I815a7d505fb7fbc384b705e5504d6861c5a67746