shellbox, shellbox-video: add support for min_avail_workers, set to 1 for video
When shellbox-video encodes videos, it blocks. This is expected
behaviour, but it means that healthchecks will connect, but never
complete on long-running encodes. Eventually workers will expire and
jobs will fail to encode. Additionally, the host *should* be unready
while it is blocking on doing an encode.

In order to address this, bump workers to 2 on shellbox-video, and use
the new min_avail_workers parameter on the healthcheck that will
validate that there is at least one free worker on the pod. During a
healthcheck on an idle pod, this will yield 1 free worker as the
healthcheck itself will block a worker.

Bug: T357309
Change-Id: Id71e6787953f53690a239c3b75796649bae8160c