cirrus: Update container version
Updates container to new version with support for providing an
auth token to the NetworkSession mediawiki auth scheme. This will
be provided with all mediawiki requests, but only if configured.
In the initial deployment the token will be unconfigured. When
configured it will need to be done from the puppet private repo.

Change-Id: I71a1537fefbad55b36894708de7ea4c0756be713