changeprop-jobqueue: reduce refreshLinks concurrency
refreshLinks *might* be causing some impact on the s4 masters - not
confirmed, but this is probably worth trying. There are spikes of
refreshLinks happening relative to the s4 issues, but there are no
real smoking guns linking the two.

If the backlog builds up even more, we might drop some events after
24h (which sometimes happens already in major buildups). This means
that some "what links here" sections might be slightly stale or
incorrect as things are moved, added or renamed.

Bug: T370304
Change-Id: Ia271f53e56a5796c039dc61ae5ac62ad6802bc86