airflow: run the initdb hook post-installation
The hook relies on configmaps, and if we run it at pre-install,
the configmaps won't be there. meaning that the job pod won't be
able to be created.

Our only option is to run it as a post-install hook. This is a bit
unfortunate, as it means that the initdb hook will run at the same
time as the scheduler. However, once the initdb job completes, the
scheduler should then be able to talk to the DB and thus start.

It's a bit awkward to my taste, but that's ok.

Change-Id: I947c827d7ff33ce5f97b863946639f7dec8f978d