airflow: run the initdb as an initcontainer
We've encountered a tricky coordination issue: we need to
run a command that requires some configmap to be mounted to
run, that should run before the webserver.

We tried a pre-install hook, but that fails because the configmaps
are not yet released at pre-install time.

We tried a post-install hook, but that fails as well, as for
the hook to be triggered, we need the webserver to be deployed
and ready, which can't happen until the DB has been initialized.

We tried running a one-off Job, which works! However, until it completes,
the webserver runs and fails in a loop.

Ultimately, running in the initdb command in an initcontainer is the
simplest form of coordination. Every time the app is deployed, we
apply the DB migrations, then start the webserver.

Bug: T372286
Change-Id: If0816b65788ecf00db874af6b282cc847c6cd2f5