cfssl-issuer: Add external-services support
As of now we hardcoded IPs for pki servers in
helmfile.d/admin_ng/cert-manager/helmfile.yaml.

cfssl-issuer does not use our scaffolding and it seemed like a quite
invasive change to add it just for this one network policy. So I decided
to create it in the chart directly.

Bug: T359423
Bug: T337928
Change-Id: I82a8de163e6032f4df68f0a2c776c13ccb683a22