k8s-controller-sidecars: adopt securityContext
There are a couple of things going on here, but the most important one
is wiring base.helper.restrictedSecurityContext into the container spec.

I've corrected package.json to the correct / current state of the world
and revendored.

This includes replacing a previously undeclared base.meta:1.0 dependency
with base.meta:2.0 (as the former now has a cursed dependency graph).

Bug: T362978
Change-Id: I2d20877855f0b80164cfaf8b947ad7cdd193ae42