airflow: perform an initial git-sync of the dags in an init container
This change ensures that the dags are pulled as part of the webserver
initialization phase.

This does _not_, however, perform regular pulls once the dags get
updated. We still need to figure out exactly how we're planning to
enable that.

Note; we perform a git sparse checkout, to only fetch wmf_airflow_common,
as well as the configured dags folder for the deployed airflow instance.
This way, we can _guarantee_ that we can only run the dags for the instance,
and nothing else.

Bug: T368033
Change-Id: Icc7475d202777b49a3e7abdd9d424513b2009d03
Depends-On: I27b73dc829393e3cdf522e2cc37bbc32eadd99e6