airflow: deploy the scheduler via a separate Deployment
We need the scheduler to be running fully separately from the webserver,
and not just as a sidecar, because this _can_ become a pretty large
resource consumer in large Airflow instances, and thus might need to be
scaled idependently.

By default, the scheduler performs processing/serialization of the DAGs
as well, but this can be carved out into a separate process if need be.

Cf https://airflow.apache.org/docs/apache-airflow/stable/core-concepts/overview.html#required-components

We rely on git sparse checkout to only checkout the airflow plugins
(as part of `wmf_airflow_common`), and clone both `wmf_airflow_common`
as well as the instance dag folder in the scheduler. Indeed, the
scheduler, running the DAG processor, needs the DAGs to serialize them
to database, whereas the webserver only needs DB access to read the DAG
code.

I took the opportunity to also bundle a couple of small changes:
- the `evalVariable` template is now recursive, so we have a value
  referencing another which references another. I ended up not needing
  it anyway, but the feature is pretty cool and might be useful.
- I enable Airflow to reach out to our MX servers, on port 25.
- The airflow and scheduler Deployments now have `component` labels.
  This means that, say, the webserver can be selected via the following
  label selector: `-l app=airflow,component=webserver`
- I also added a `.Values.monitoring.statsd` block that was missing to
  deploy the statsd exporter sidecar.

Bug: T368737
Change-Id: Ifda244253217f591fa19f6a5ee21fe2ce742cde9